+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/Payload.java
@@ -1,5 +1,6 @@
 package com.walmartlabs.concord.server.process;
 
+import com.walmartlabs.concord.project.model.ProjectDefinition;
 import com.walmartlabs.concord.server.process.keys.AttachmentKey;
 import com.walmartlabs.concord.server.process.keys.HeaderKey;
 
@@ -17,6 +18,8 @@ public class Payload {
     public static final HeaderKey<Path> WORKSPACE_DIR = HeaderKey.register("_workspace", Path.class);
     public static final HeaderKey<Map> REQUEST_DATA_MAP = HeaderKey.register("_meta", Map.class);
     public static final HeaderKey<String> RESUME_EVENT_NAME = HeaderKey.register("_resumeEventName", String.class);
+    public static final HeaderKey<String[]> ACTIVE_PROFILES = HeaderKey.register("_activeProfiles", String[].class);
+    public static final HeaderKey<ProjectDefinition> PROJECT_DEFINITION = HeaderKey.register("_projectDef", ProjectDefinition.class);
 
     public static final AttachmentKey WORKSPACE_ARCHIVE = AttachmentKey.register("_workspaceArchive");