+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 ### Added
 
 - support for `.concord.yml` files. Those files can contain flow and
-form definitions, default variables and "profiles".
+form definitions, default variables and "profiles";
+- allow overriding`dependencies` in a project file, defaults or user
+requests.
 
 ### Breaking